+++ src/renderer/src/components/sider/tun-switcher.tsx
@@ -3,19 +3,16 @@ import { useControledMihomoConfig } from '@renderer/hooks/use-controled-mihomo-c
 import BorderSwitch from '@renderer/components/base/border-swtich'
 import { TbDeviceIpadHorizontalBolt } from 'react-icons/tb'
 import { useLocation } from 'react-router-dom'
-import { encryptString, isEncryptionAvailable, restartCore } from '@renderer/utils/ipc'
+import { restartCore } from '@renderer/utils/ipc'
 import { useSortable } from '@dnd-kit/sortable'
 import { CSS } from '@dnd-kit/utilities'
-import { platform } from '@renderer/utils/init'
-import React, { useState } from 'react'
+import React from 'react'
 import { useAppConfig } from '@renderer/hooks/use-app-config'
-import BasePasswordModal from '../base/base-password-modal'
 
 const TunSwitcher: React.FC = () => {
   const location = useLocation()
   const match = location.pathname.includes('/tun') || false
-  const [openPasswordModal, setOpenPasswordModal] = useState(false)
-  const { appConfig, patchAppConfig } = useAppConfig()
+  const { appConfig } = useAppConfig()
   const { tunCardStatus = 'col-span-1' } = appConfig || {}
   const { controledMihomoConfig, patchControledMihomoConfig } = useControledMihomoConfig()
   const { tun } = controledMihomoConfig || {}
@@ -32,19 +29,6 @@ const TunSwitcher: React.FC = () => {
   })
   const transform = tf ? { x: tf.x, y: tf.y, scaleX: 1, scaleY: 1 } : null
   const onChange = async (enable: boolean): Promise<void> => {
-    if (enable && platform === 'darwin') {
-      const encryptionAvailable = await isEncryptionAvailable()
-      if (!appConfig?.encryptedPassword && encryptionAvailable) {
-        setOpenPasswordModal(true)
-        return
-      }
-      if (!appConfig?.encryptedPassword && !encryptionAvailable) {
-        alert('加密不可用，请手动给内核授权')
-        await patchAppConfig({ encryptedPassword: [] })
-        return
-      }
-    }
-
     if (enable) {
       await patchControledMihomoConfig({ tun: { enable }, dns: { enable: true } })
     } else {
@@ -65,24 +49,6 @@ const TunSwitcher: React.FC = () => {
       }}
       className={`${tunCardStatus} tun-card`}
     >
-      {openPasswordModal && (
-        <BasePasswordModal
-          onCancel={() => setOpenPasswordModal(false)}
-          onConfirm={async (password: string) => {
-            try {
-              const encrypted = await encryptString(password)
-              await patchAppConfig({ encryptedPassword: encrypted })
-              await patchControledMihomoConfig({ tun: { enable: true }, dns: { enable: true } })
-              await restartCore()
-              window.electron.ipcRenderer.send('updateTrayMenu')
-              setOpenPasswordModal(false)
-            } catch (e) {
-              alert(e)
-            }
-          }}
-        />
-      )}
-
       <Card
         fullWidth
         ref={setNodeRef}