+++ src/renderer/src/pages/mihomo.tsx
@@ -89,7 +89,7 @@ const Mihomo: React.FC = () => {
                     setTimeout(() => {
                       PubSub.publish('mihomo-core-changed')
                     }, 2000)
-                    if (platform === 'linux') {
+                    if (platform !== 'win32') {
                       new Notification('内核权限丢失', {
                         body: '内核升级成功，若要使用虚拟网卡（Tun），请到虚拟网卡页面重新手动授权内核'
                       })