+++ changelog.md
@@ -1,3 +1,11 @@
+### Breaking Changes
+
+- macOS 改用 pkg 安装方式，不再支持 dmg 安装方式，因此本次更新需要手动下载安装包进行安装
+
 ### New Features
 
-- Linux 不再存储 root 密码
+- macOS/Linux 均不再存储 root 密码
+
+### Bug Fixes
+
+- 修复 macOS 10.15 无法安装的问题